Fix PEM key format
It is incorrect to write the PEM file with a "EC PRIVATE KEY" because
it could be RSA as well (or any other algorithm).

Besides, the format specified by rfc5915 does not include an
`privateKeyAlgorithm` field. But, the data sent by the SPIFFE workload
API is in the `PrivateKeyInfo` format specified by
https://www.rfc-editor.org/rfc/rfc5208#section-5. Thus strict parsers
- such as the one used by erlang/OTP - reject the key.